back-end/requirements.txt
fastapi==0.101.0
uvicorn==0.25.0
pandas==2.1.0
joblib==1.3.2
python-jose==3.3.0
cryptography==41.0.3
pydantic==2.7.0
scikit-learn==1.5.1
httpx==0.28.1
pytest==8.3.5
